uploadphoto refactored
