+++ services/newuser.js
@@ -131,6 +131,41 @@ export const updateArea = (id, name, token, geostore) => {
     });
 };
 
+export const uploadPhoto = (file, user) => {
+  return new Promise((resolve, reject) => {
+    const reader = new FileReader();
+    reader.readAsDataURL(file);
+
+    reader.onload = () => {
+      const bodyObj = {
+        data: {
+          attributes: {
+            user_id: user.id,
+            avatar: reader.result
+          }
+        }
+      };
+
+      return fetch(`${process.env.WRI_API_URL}/profile`, {
+        method: 'POST',
+        body: JSON.stringify(bodyObj),
+        headers: {
+          'Content-Type': 'application/json',
+          Authorization: user.token
+        }
+      })
+        .then(response => response.json())
+        .then(({ data }) => {
+          resolve(data.attributes.avatar.original);
+        });
+    };
+
+    reader.onerror = (error) => {
+      reject(error);
+    };
+  });
+}
+
 
 export default {
   loginUser,
@@ -140,5 +175,6 @@ export default {
   getUserAreas,
   deleteArea,
   createArea,
-  updateArea
+  updateArea,
+  uploadPhoto
 };
