+++ services/user.js
@@ -116,38 +116,5 @@ export default class UserService {
       .then(response => response.json());
   }
 
-  uploadPhoto(file, user) {
-    return new Promise((resolve, reject) => {
-      const reader = new FileReader();
-      reader.readAsDataURL(file);
 
-      reader.onload = () => {
-        const bodyObj = {
-          data: {
-            attributes: {
-              user_id: user.id,
-              avatar: reader.result
-            }
-          }
-        };
-
-        return fetch(`${process.env.WRI_API_URL}/profile`, {
-          method: 'POST',
-          body: JSON.stringify(bodyObj),
-          headers: {
-            'Content-Type': 'application/json',
-            Authorization: user.token
-          }
-        })
-          .then(response => response.json())
-          .then(({ data }) => {
-            resolve(data.attributes.avatar.original);
-          });
-      };
-
-      reader.onerror = (error) => {
-        reject(error);
-      };
-    });
-  }
 }
